delete/refactor : #35 - payment 모듈 제거>reservation으로 합병
- CONSUMER_SERVER_PORT -> SERVER_PORT
