+++ escape-consumer/src/main/resources/application.yml
@@ -37,4 +37,4 @@ jwt:
   secret:
     key: ${JWT_SECRET_KEY}
 server:
-  port: ${CONSUMER_SERVER_PORT}
+  port: ${SERVER_PORT}
