+++ escape-core/src/main/java/com/sparta/domain/reservation/entity/Reservation.java
@@ -8,6 +8,7 @@ import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Getter
@@ -27,6 +28,7 @@ public class Reservation extends TimeStamped {
     @Column(nullable = false)
     private PaymentStatus paymentStatus;
 
+    @Setter
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private ReservationStatus reservationStatus;
