--- escape-payment/src/main/resources/application.yml
@@ -1,53 +1,0 @@
-spring:
-  datasource:
-    url: ${DATASOURCE_URL}
-    username: ${DATASOURCE_USERNAME}
-    password: ${DATASOURCE_PASSWORD}
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        show_sql: true
-        format_sql: true
-        use_sql_comments: true
-  mail:
-    host: ${MAIL_HOST} #smtp 서버 주소
-    port: ${MAIL_PORT} # 메일 인증서버 포트
-    username: ${MAIL_USERNAME} #네이버 아이디
-    password: ${MAIL_PASSWORD} #네이버 비밀번호
-    properties:
-      mail:
-        smtp:
-          starttls:
-            enable: true
-            required: true
-          auth: true
-          connectiontimeout: 5000
-          timeout: 5000
-          writetimeout: 50000
-          ssl:
-            trust: smtp.gmail.com
-            enable: true #반드시 필요 -> SSL 필요로 되어있으면
-  data:
-    redis:
-      port: ${REDIS_MAIL_PORT}
-      host: ${REDIS_MAIL_HOST}
-jwt:
-  secret:
-    key: ${JWT_SECRET_KEY}
-server:
-  port: ${SERVER_PORT}
-
-kakao-payment:
-  admin-key: ${PAY_ADMIN_KEY}
-  cid: TC0ONETIME
-  approve-url:  http://localhost:8089/payment/kakaoPaySuccess
-  cancel-url: http://localhost:8089/payment/kakaoPayCancel
-  fail-url: http://localhost:8089/payment/kakaoPayFail
-
-toss:
-  payment:
-    client-key : ${CLIENT-KEY}
-    secret-key : ${SECRET-KEY}
