+++ escape-reservation/src/main/java/com/sparta/domain/kakaopayment/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.kakaopayment.repository;
+package com.sparta.domain.kakaopayment.repository;
 
-import com.sparta.kakaopayment.entity.KakaoPayment;
+import com.sparta.domain.kakaopayment.entity.KakaoPayment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<KakaoPayment, Integer> {
