+++ escape-reservation/src/main/java/com/sparta/domain/reservation/dto/GetReservationResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.dto;
+package com.sparta.domain.reservation.dto;
 
 import com.sparta.domain.reservation.entity.PaymentStatus;
 import com.sparta.domain.reservation.entity.Reservation;
