+++ escape-reservation/src/main/resources/application.yml
@@ -15,8 +15,8 @@ spring:
   mail:
     host: ${MAIL_HOST} #smtp 서버 주소
     port: ${MAIL_PORT} # 메일 인증서버 포트
-    username: ${MAIL_USERNAME} #구글 아이디
-    password: ${MAIL_PASSWORD} #구글 비밀번호
+    username: ${MAIL_USERNAME} #네이버 아이디
+    password: ${MAIL_PASSWORD} #네이버 비밀번호
     properties:
       mail:
         smtp:
@@ -39,3 +39,15 @@ jwt:
     key: ${JWT_SECRET_KEY}
 server:
   port: ${SERVER_PORT}
+
+kakao-payment:
+  admin-key: ${PAY_ADMIN_KEY}
+  cid: TC0ONETIME
+  approve-url:  http://localhost:${SERVER_PORT}/payment/kakaoPaySuccess
+  cancel-url: http://localhost:${SERVER_PORT}/payment/kakaoPayCancel
+  fail-url: http://localhost:${SERVER_PORT}/payment/kakaoPayFail
+
+toss:
+  payment:
+    client-key : ${CLIENT-KEY}
+    secret-key : ${SECRET-KEY}
