+++ settings.gradle
@@ -8,5 +8,4 @@ include 'escape-follow'
 include 'escape-store'
 include 'escape-review'
 include 'escape-search'
-include 'escape-payment'
